+++ plugins/events/get.js
@@ -3,23 +3,23 @@
 const boom = require('boom');
 const joi = require('joi');
 const schema = require('screwdriver-data-schema');
-const getSchema = schema.models.build.get;
-const idSchema = joi.reach(schema.models.build.base, 'id');
+const getSchema = schema.models.event.get;
+const idSchema = joi.reach(schema.models.event.base, 'id');
 
 module.exports = () => ({
     method: 'GET',
-    path: '/builds/{id}',
+    path: '/events/{id}',
     config: {
-        description: 'Get a single build',
-        notes: 'Returns a build record',
-        tags: ['api', 'builds'],
+        description: 'Get a single event',
+        notes: 'Returns a event record',
+        tags: ['api', 'events'],
         handler: (request, reply) => {
-            const factory = request.server.app.buildFactory;
+            const factory = request.server.app.eventFactory;
 
             return factory.get(request.params.id)
                 .then((model) => {
                     if (!model) {
-                        throw boom.notFound('Build does not exist');
+                        throw boom.notFound('Event does not exist');
                     }
 
                     return reply(model.toJson());