+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 Features:
   * Create an event when creating a build
   * Add eventId to the build
+  * Add `GET /events/{id}` and `GET /events/{id}/builds` endpoints
 
 ## 3.3.0